feature/IVYPORTAL-18677-Third-party-application-with-permissions - Update Migration
